+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next
+
+- **[Breaking change]** Update to `swf-types@0.11.0`.
+- **[Feature]** Implement emitter for `DoAbc`.
+
 # 0.10.0 (2020-01-20)
 
 - **[Breaking change]** Update to `swf-types@0.10.0`.